+++ sqlalchemy_json_api/__init__.py
@@ -6,5 +6,6 @@ from .exc import (  # noqa
     UnknownModel
 )
 from .query_builder import QueryBuilder, RESERVED_KEYWORDS  # noqa
+from .utils import assert_json_document  # noqa
 
 __version__ = '0.4.2'